refactor: move page file folder to src
