+++ src/containers/Home/components/Announcement.tsx
@@ -2,7 +2,7 @@ import React, { FC } from 'react'
 import styled from 'styled-components'
 import { SVG_SPRITE } from 'src/shared/constants'
 import { flexMixin } from 'src/styled/mixins'
-import svgIcons from 'src/static/svg-sprite.svg'
+import svgIcons from 'src/static/images/svg-sprite.svg'
 import { IAnnouncement } from '../types'
 
 interface Props {
