+++ src/static/images/normal.cur
